feat(cli): make `key convert` subcommand user friendly
