+++ NineChronicles.Headless.Executable/Commands/KeyCommands.cs
@@ -1,4 +1,4 @@
-using NineChronicles.Headless.Executable.Commands;
+using NineChronicles.Headless.Executable.Commands.Key;
 
 #nullable enable
 // Copied from https://git.io/Jqc0q
